backend/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncpg==0.29.0
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
fastapi==0.110.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
kombu==5.3.5
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic_core==2.16.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
redis==5.0.2
requests==2.31.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.27.1
vine==5.1.0
wcwidth==0.2.13
